backend/requirements.txt
fastapi
uvicorn
google-generativeai
httpx
pydantic
python-dotenv
pytest
